Added the user_database which stores all the followed users. * The profile picture on the home screen follow tab will use the same base 64 method as the thumbnails to load. * The thumbnailList in IsostaUser.kt was made nullable.  Not null assertion operators (git status) were used when calling the thumbnailList.  The thumbnailList should only be null when loading the user from the database.

+++ app/src/main/java/io/github/azusalad/isosta/model/Thumbnail.kt
@@ -10,8 +10,8 @@ data class Thumbnail (
     val picture: String,
     val text: String,
     val sourceHandle: String,
-    val date: Int = 0,
+    val date: Long = 0,
 
     //@ColumnInfo(typeAffinity = ColumnInfo.BLOB)
-    var pictureString: String? = null
+    var pictureString: String? = null  // Thumbnail picture as Base64
 )
